utils.typing, configs: implement pydantic validators for port, ip and http url
