+++ otaclient/_utils/typing.py
@@ -14,9 +14,11 @@
 
 
 from __future__ import annotations
+import ipaddress
 from pathlib import Path
 from typing import Any, Callable, TypeVar, Union
-from typing_extensions import Concatenate, ParamSpec
+from typing_extensions import Annotated, Concatenate, ParamSpec
+from urllib.parse import urlparse
 
 P = ParamSpec("P")
 T = TypeVar("T")
@@ -51,3 +53,40 @@ def copy_callable_typehint_to_method(_source: Callable[P, Any]):
 
 def check_port(_in: Any) -> bool:
     return isinstance(_in, int) and _in >= 0 and _in <= 65535
+
+
+# pydantic helpers
+
+
+def gen_strenum_validator(enum_type: type[T]) -> Callable[[T | str | Any], T]:
+    """A before validator generator that converts input value into enum
+    before passing it to pydantic validator.
+
+    NOTE(20240129): as upto pydantic v2.5.3, (str, Enum) field cannot
+                    pass strict validation if input is str.
+    """
+
+    def _inner(value: T | str | Any) -> T:
+        assert isinstance(
+            value, (enum_type, str)
+        ), f"{value=} should be {enum_type} or str type"
+        return enum_type(value)
+
+    return _inner
+
+
+def port_validator(value: int) -> int:
+    assert value >= 0 and value <= 65535, f"invalid port number: {value}"
+    return value
+
+
+def ip_validator(value: str) -> str:
+    ipaddress.ip_address(value)
+    return value
+
+
+def http_url_validator(value: str) -> str:
+    _parsed = urlparse(value)
+    assert _parsed.scheme in ("http", "https"), f"{value} is not a http(s) URL"
+    assert _parsed.netloc, f"{value} doesn't contain netloc"
+    return value
